Refactor vacancyService test with teardown inject of service
Eslint for unused vars

Error handling vacancyService

+++ src/wwwroot/app/services/vacancyService.js
@@ -26,7 +26,11 @@ export default class VacancyService {
    }
 
    deleteVacancy(entity) {
-      var additionalUrl = VACANCY_URL + entity.Id;
-      _HttpService.remove(additionalUrl, entity);
+      if (entity.Id === undefined) {
+         throw new Error('Id should be specified');
+      } else {
+         var additionalUrl = VACANCY_URL + entity.Id;
+         _HttpService.remove(additionalUrl, entity);
+      }
    }
 }
